refactor(mark): remove media and author references from Mark and MarkOccurrence; rename services to CommandService; update mappers and DTOs accordingly
