+++ mark/src/main/java/pt/estga/mark/dtos/MarkOccurrenceDto.java
@@ -1,16 +1,12 @@
 package pt.estga.mark.dtos;
 
 import java.time.Instant;
-import java.util.UUID;
 
 public record MarkOccurrenceDto(
     Long id,
     Long markId,
     Long monumentId,
     MarkDto mark,
-    UUID coverId,
-    Long authorId,
-    String authorName,
     Instant publishedAt,
     Boolean active
 ) { }
